+++ src/configuration/configuration.cpp
@@ -30,10 +30,10 @@ void PrintConfiguration(){
   LOG_INFO("%30s", "//===-------------- PELOTON CONFIGURATION --------------===//");
   LOG_INFO(" ");
 
-  LOG_INFO("%30s: %10lu", "Port", FLAGS_port);
+  LOG_INFO("%30s: %10llu", "Port", (unsigned long long) FLAGS_port);
   LOG_INFO("%30s: %10s",  "Socket Family", FLAGS_socket_family.c_str());
   LOG_INFO("%30s: %10s", "Statistics", FLAGS_stats_mode ? "enabled" : "disabled");
-  LOG_INFO("%30s: %10lu", "Max Connections", FLAGS_max_connections);
+  LOG_INFO("%30s: %10llu", "Max Connections", (unsigned long long) FLAGS_max_connections);
   LOG_INFO("%30s: %10s", "Index Tuner", FLAGS_index_tuner ? "enabled" : "disabled");
   LOG_INFO("%30s: %10s", "Layout Tuner", FLAGS_layout_tuner ? "enabled" : "disabled");
   LOG_INFO("%30s: %10s",  "Code-generation", FLAGS_codegen ? "enabled" : "disabled");
