+++ test/sql/order_by_sql_test.cpp
@@ -88,7 +88,7 @@ TEST_F(OrderBySQLTests, PerformanceTest) {
 
   LOG_INFO(
       "OrderBy Query (table size:%d) with Limit 10 Execution Time is: %lu ms",
-      table_size, latency);
+      table_size, (unsigned long) latency);
 
   // test OrderBy without Limit
   start_time = std::chrono::system_clock::now();
@@ -103,7 +103,7 @@ TEST_F(OrderBySQLTests, PerformanceTest) {
                 end_time - start_time).count();
 
   LOG_INFO("OrderBy Query (table size:%d) Execution Time is: %lu ms",
-           table_size, latency);
+           table_size, (unsigned long) latency);
 
   // free the database just created
   txn = txn_manager.BeginTransaction();
